修复部分问题，添加测试结果
