+++ hlp/chat/common/attention.py
@@ -1,5 +1,6 @@
 import tensorflow as tf
 
+
 def scaled_dot_product_attention(q, k, v, mask):
     matmul_qk = tf.matmul(q, k, transpose_b=True)
     dk = tf.cast(tf.shape(k)[-1], tf.float32)
@@ -58,7 +59,7 @@ class MultiHeadAttention(tf.keras.layers.Layer):
         return output, attention_weights
 
 
-class BahdanauAttention(tf.keras.Model):
+class BahdanauAttention(tf.keras.layers.Layer):
     def __init__(self, units):
         super(BahdanauAttention, self).__init__()
         self.W1 = tf.keras.layers.Dense(units)
@@ -68,11 +69,8 @@ class BahdanauAttention(tf.keras.Model):
     def call(self, query, values):
         hidden_with_time_axis = tf.expand_dims(query, 1)
         score = self.V(tf.nn.tanh(
-            self.W1(values) + self.W2(hidden_with_time_axis)
-        ))
-
+            self.W1(values) + self.W2(hidden_with_time_axis)))
         attention_weights = tf.nn.softmax(score, axis=1)
-
         context_vector = attention_weights * values
         context_vector = tf.reduce_sum(context_vector, axis=1)
