+++ hlp/chat/config/seq2seq.json
@@ -1,14 +1,15 @@
 {
   "mode": "train",
   "train_data": "checkpoint\\seq2seq",
-  "resource_data": "data\\text.txt",
-  "tokenized_data": "data\\text_tokenized.txt",
+  "resource_data": "data\\train.txt",
+  "tokenized_data": "data\\train_tokenized.txt",
   "start": "E",
   "end": "M",
   "enc_vocab_size": 20000,
   "dec_vocab_size": 20000,
-  "embedding_dim": 128,
-  "layer_size": 256,
-  "max_train_data_size": 50000,
-  "batch_size": 128
+  "embedding_dim": 256,
+  "layer_size": 1024,
+  "max_train_data_size": 100,
+  "batch_size": 64,
+  "epochs": 100
 }
