--- hlp/chat/data/text.txt
@@ -1,96 +1,0 @@
-谢谢你所做的一切
-你开心就好
-开心
-嗯因为你的心里只有学习
-某某某，还有你
-这个某某某用的好
-
-你们宿舍都是这么厉害的人吗
-眼睛特别搞笑这土也不好捏但就是觉得挺可爱
-特别可爱啊
-
-今天好点了吗？
-一天比一天严重
-吃药不管用，去打一针。别拖着
-
-是的。下辈子想做只萤火虫
-可是萤火虫太容易被抓了还是改一个吧
-不，我只想奋不顾身扑火
-
-加油，三月动起来，五月笑起来
-正解你为什么就那么厉害呢
-哈哈，没办法，智商就是这么高
-你这是要开始得瑟了吗！好啦！你最厉害！
-哈哈哈哈
-
-好身材，秀出来
-哈哈哈其实我是胖的
-不会的
-
-因为我网络差吗，加载不出来啊
-是什么网络啊，移动的可能比较慢
-啊真的是我自家网络的问题，用别家的就好了
-
-这是克拉玛依咩！
-对啊！有没有啥推荐的快快快
-哈哈哈哈哈，要吃东西的话就去那个汉博，那边比较多克拉玛依的凉皮也可以尝尝然后可以去九龙潭看看，克拉玛依图书馆啊，一号井啊都是克拉玛依旅游比较常去的地方
-良心评论
-
-这个側颜可以
-这个繁体很厉害
-为了配你的颜
-哥，
-
-我今天腿都废了，你们过节，我搬砖
-辛苦啊，圣诞节还去赚大钱了加油
-毕竟是没男朋友的人，什么节都是一样的
-
-保留几分
-晾毛巾晾毛巾叫你天天晾毛巾
-看看几点了还不快去吃饭
-
-经常都记着出门要带学生证，每次都忘
-我刚检完票候车就发现了但是来不及了
-尴尬
-
-嘉宁
-胸不错
-摸够了吧。
-小小贡献不足为提
-
-我也看过了
-握个手吧
-好的勉强同意了
-
-哈哈哈哈哈哈哈哈哈哈哈哈
-你哈哈个啥
-别人晒得都是某某某搞事情为啥只有你是瘦了！太尼玛懂你了笑哭我了
-这个太假了
-
-南通的海风真的是扛不住
-还好啦，挺惬意的，园博园的海风才舒坦，有兴趣可以去园博园吹吹海风！
-滨江公园的风已经很可以了
-
-看着就难吃
-又不是你吃，莫操心
-心疼你
-滚蛋
-不滚
-
-外出了？
-嗯啊
-好好玩，注意安全哈
-
-你中招了？
-差不多！
-那么大人，还是要注意点。
-
-我也预定了重庆站！月的
-可是号已经售光了啊我打算买号的
-号显示的缺货，然后我缺货登记了两张。号的呢？啥时候开始售？
-哈哈哈哈原来是这样啊我用的软件只显示了马上开售不知道撒时候售啊太揪心号的就没有买到号的一定要买到啊
-好伤心啊，你是什么软件？我都不知道在哪儿买
-
-男票哈
-放荡不羁爱自由的我哪里来的男朋友呢
-哈哈哈还是酒太好喝手机太好玩的过
