--- hlp/chat/data/text_tokenized.txt
@@ -1,55 +1,0 @@
-谢谢 你 所 做 的 一切	你 开心 就 好
-你 开心 就 好	开心
-开心	嗯 因为 你 的 心里 只有 学习
-嗯 因为 你 的 心里 只有 学习	某某某 还有 你
-某某某 还有 你	这个 某某某 用 的 好
-你们 宿舍 都 是 这么 厉害 的 人 吗	眼睛 特别 搞笑 这土 也 不好 捏 但 就是 觉得 挺 可爱
-眼睛 特别 搞笑 这土 也 不好 捏 但 就是 觉得 挺 可爱	特别 可爱 啊
-今天 好点 了 吗	一天 比 一天 严重
-一天 比 一天 严重	吃药 不管 用 去 打 一针 别 拖 着
-是 的 下辈子 想 做 只 萤火虫	可是 萤火虫 太 容易 被 抓 了 还是 改 一个 吧
-可是 萤火虫 太 容易 被 抓 了 还是 改 一个 吧	不 我 只 想 奋不顾身 扑火
-加油 三月 动 起来 五月 笑 起来	正解 你 为什么 就 那么 厉害 呢
-正解 你 为什么 就 那么 厉害 呢	哈哈 没 办法 智商 就是 这么 高
-哈哈 没 办法 智商 就是 这么 高	你 这 是 要 开始 得瑟 了 吗 好 啦 你 最 厉害
-你 这 是 要 开始 得瑟 了 吗 好 啦 你 最 厉害	哈哈哈哈
-好身材 秀 出来	哈哈哈 其实 我 是 胖 的
-哈哈哈 其实 我 是 胖 的	不会 的
-因为 我 网络 差 吗 加载 不 出来 啊	是 什么 网络 啊 移动 的 可能 比较慢
-是 什么 网络 啊 移动 的 可能 比较慢	啊 真的 是 我 自家 网络 的 问题 用 别家 的 就 好 了
-这是 克拉玛依 咩	对 啊 有没有 啥 推荐 的 快快 快
-对 啊 有没有 啥 推荐 的 快快 快	哈哈哈 哈哈 要 吃 东西 的话 就 去 那个 汉博 那边 比较 多 克拉玛依 的 凉皮 也 可以 尝尝 然后 可以 去 九 龙潭 看看 克拉玛依 图书馆 啊 一号 井 啊 都 是 克拉玛依 旅游 比较 常去 的 地方
-哈哈哈 哈哈 要 吃 东西 的话 就 去 那个 汉博 那边 比较 多 克拉玛依 的 凉皮 也 可以 尝尝 然后 可以 去 九 龙潭 看看 克拉玛依 图书馆 啊 一号 井 啊 都 是 克拉玛依 旅游 比较 常去 的 地方	良心 评论
-这个 側颜 可以	这个 繁体 很 厉害
-这个 繁体 很 厉害	为了 配 你 的 颜
-为了 配 你 的 颜	哥
-我 今天 腿 都 废 了 你们 过节 我 搬 砖	辛苦 啊 圣诞节 还 去 赚大钱 了 加油
-辛苦 啊 圣诞节 还 去 赚大钱 了 加油	毕竟 是 没 男朋友 的 人 什么 节 都 是 一样 的
-保留 几分	晾 毛巾 晾 毛巾 叫 你 天天 晾 毛巾
-晾 毛巾 晾 毛巾 叫 你 天天 晾 毛巾	看看 几点 了 还 不快 去 吃饭
-经常 都 记着 出门 要 带 学生证 每次 都 忘	我 刚检 完票 候车 就 发现 了 但是 来不及 了
-我 刚检 完票 候车 就 发现 了 但是 来不及 了	尴尬
-嘉宁	胸 不错
-胸 不错	摸够 了 吧
-摸够 了 吧	小小 贡献 不足 为 提
-我 也 看过 了	握个 手 吧
-握个 手 吧	好 的 勉强 同意 了
-哈哈哈 哈哈哈 哈哈哈 哈哈哈	你 哈哈 个 啥
-你 哈哈 个 啥	别人 晒 得 都 是 某某某 搞 事情 为啥 只有 你 是 瘦 了 太 尼玛 懂 你 了 笑 哭 我 了
-别人 晒 得 都 是 某某某 搞 事情 为啥 只有 你 是 瘦 了 太 尼玛 懂 你 了 笑 哭 我 了	这个 太假 了
-南通 的 海风 真的 是 扛不住	还好 啦 挺 惬意 的 园 博园 的 海风 才 舒坦 有 兴趣 可以 去园 博园 吹 吹 海风
-还好 啦 挺 惬意 的 园 博园 的 海风 才 舒坦 有 兴趣 可以 去园 博园 吹 吹 海风	滨江 公园 的 风 已经 很 可以 了
-看着 就 难吃	又 不是 你 吃 莫 操心
-又 不是 你 吃 莫 操心	心疼 你
-心疼 你	滚蛋
-滚蛋	不滚
-外出 了	嗯 啊
-嗯 啊	好 好玩 注意安全 哈
-你 中招 了	差不多
-差不多	那么 大人 还是 要 注意 点
-我 也 预定 了 重庆 站 月 的	可是 号 已经 售光 了 啊 我 打算 买 号 的
-可是 号 已经 售光 了 啊 我 打算 买 号 的	号 显示 的 缺货 然后 我 缺货 登记 了 两张 号 的 呢 啥时候 开始 售
-号 显示 的 缺货 然后 我 缺货 登记 了 两张 号 的 呢 啥时候 开始 售	哈哈哈哈 原来 是 这样 啊 我 用 的 软件 只 显示 了 马上 开售 不 知道 撒 时候 售 啊 太 揪心 号 的 就 没有 买到 号 的 一定 要 买 到 啊
-哈哈哈哈 原来 是 这样 啊 我 用 的 软件 只 显示 了 马上 开售 不 知道 撒 时候 售 啊 太 揪心 号 的 就 没有 买到 号 的 一定 要 买 到 啊	好 伤心 啊 你 是 什么 软件 我 都 不 知道 在 哪儿 买
-男票 哈	放荡不羁 爱 自由 的 我 哪里 来 的 男朋友 呢
-放荡不羁 爱 自由 的 我 哪里 来 的 男朋友 呢	哈哈哈 还是 酒太 好喝 手机 太 好玩 的 过
