--- hlp/chat/model/Seq2Seq/model.py
@@ -1,94 +1,0 @@
-import tensorflow as tf
-import config.getConfig as gConfig
-import common.attention as attention
-
-
-class Encoder(tf.keras.Model):
-    def __init__(self, vocab_size, embedding_dim, enc_units, batch_sz):
-        super(Encoder, self).__init__()
-        self.batch_sz = batch_sz
-        self.enc_units = enc_units
-        self.embedding = tf.keras.layers.Embedding(vocab_size, embedding_dim)
-        self.gru = tf.keras.layers.GRU(self.enc_units, return_sequences=True, return_state=True,
-                                       recurrent_initializer='glorot_uniform')
-
-    def call(self, x, hidden):
-        x = self.embedding(x)
-        output, state = self.gru(x, initial_state=hidden)
-        return output, state
-
-    def initialize_hidden_state(self):
-        return tf.zeros((self.batch_sz, self.enc_units))
-
-
-class Decoder(tf.keras.Model):
-    def __init__(self, vocab_size, embedding_dim, dec_units, batch_sz):
-        super(Decoder, self).__init__()
-        self.batch_sz = batch_sz
-        self.dec_units = dec_units
-        self.embedding = tf.keras.layers.Embedding(vocab_size, embedding_dim)
-        self.gru = tf.keras.layers.GRU(self.dec_units,
-                                       return_sequences=True,
-                                       return_state=True,
-                                       recurrent_initializer='glorot_uniform')
-        self.fc = tf.keras.layers.Dense(vocab_size)
-
-        self.attention = attention.BahdanauAttention(self.dec_units)
-
-    def call(self, x, hidden, enc_output):
-        context_vector, attention_weights = self.attention(hidden, enc_output)
-        x = self.embedding(x)
-        x = tf.concat([tf.expand_dims(context_vector, 1), x], axis=-1)
-        output, state = self.gru(x)
-        output = tf.reshape(output, (-1, output.shape[2]))
-        x = self.fc(output)
-
-        return x, state, attention_weights
-
-
-encoder = Encoder(gConfig.vocab_inp_size, gConfig.embedding_dim, gConfig.units, gConfig.BATCH_SIZE)
-decoder = Decoder(gConfig.vocab_tar_size, gConfig.embedding_dim, gConfig.units, gConfig.BATCH_SIZE)
-
-optimizer = tf.keras.optimizers.Adam()
-loss_object = tf.keras.losses.SparseCategoricalCrossentropy(from_logits=True)
-
-
-def loss_function(real, pred):
-    # 这里进来的real和pred的shape为（128,）
-    mask = tf.math.logical_not(tf.math.equal(real, 0))
-    loss_ = loss_object(real, pred)
-
-    # 这里要注意了，因为前面我们对于短的句子进行了填充，所以对于填充的部分，我们不能用于计算损失
-    mask = tf.cast(mask, dtype=loss_.dtype)
-    loss_ *= mask
-
-    return tf.reduce_mean(loss_)
-
-
-checkpoint = tf.train.Checkpoint(optimizer=optimizer, encoder=encoder, decoder=decoder)
-
-
-# @tf.function
-def train_step(inp, targ, targ_lang, enc_hidden):
-    loss = 0
-
-    with tf.GradientTape() as tape:
-        enc_output, enc_hidden = encoder(inp, enc_hidden)
-        dec_hidden = enc_hidden
-
-        # 这里初始化decoder的输入，首个token为start，shape为（128, 1）
-        dec_input = tf.expand_dims([targ_lang.word_index['start']] * gConfig.BATCH_SIZE, 1)
-
-        # 这里针对每个训练出来的结果进行损失计算
-        for t in range(1, targ.shape[1]):
-            predictions, dec_hidden, _ = decoder(dec_input, dec_hidden, enc_output)
-            loss += loss_function(targ[:, t], predictions)
-            dec_input = tf.expand_dims(targ[:, t], 1)
-
-    batch_loss = (loss / int(targ.shape[1]))
-    variables = encoder.trainable_variables + decoder.trainable_variables
-    gradients = tape.gradient(loss, variables)
-
-    optimizer.apply_gradients(zip(gradients, variables))
-
-    return batch_loss
